Updated Server File
Can get returns from database for one bird (unsure how we will implement the search feature). Also could not get the json parser to work, so that is something to look at in class time. Also how will we be getting the latitude and longitude for the birds? Will it be stored in the database?